Refactor tools-file
